Add a web folder w/ webpack config.
