+++ .gitignore
@@ -1,4 +1,5 @@
 node_modules/
 dist/
+web/bundle.js
 .tenant-interview-state.json
 tenant.schema.json
