+++ package.json
@@ -8,7 +8,9 @@
     "schema": "typescript-json-schema tsconfig.json Tenant > tenant.schema.json",
     "build": "tsc",
     "watch": "tsc --watch",
-    "start": "node dist/main.js"
+    "start": "node dist/main.js",
+    "build-web": "webpack",
+    "watch-web": "webpack --watch"
   },
   "author": "",
   "license": "ISC",
@@ -17,7 +19,12 @@
     "@types/node": "^10.5.2",
     "chalk": "^2.4.1",
     "minimist": "^1.2.0",
+    "ts-loader": "^4.4.2",
     "typescript": "^2.9.2",
-    "typescript-json-schema": "^0.26.0"
+    "typescript-json-schema": "^0.26.0",
+    "webpack": "^4.16.1"
+  },
+  "devDependencies": {
+    "webpack-cli": "^3.0.8"
   }
 }
