+++ web/index.html
@@ -1,0 +1,5 @@
+<!DOCTYPE html>
+<meta charset="utf-8">
+<title>Interview fun</title>
+<h1>Interview fun</h1>
+<script src="bundle.js"></script>
