+++ web/index.ts
@@ -1,0 +1,3 @@
+import { addDays } from '../lib/util';
+
+console.log('a week from today is', addDays(new Date(), 7));
